+++ delivery/src/main/java/com/nowayback/delivery/domain/delivery/vo/HubId.java
@@ -1,7 +1,7 @@
 package com.nowayback.delivery.domain.delivery.vo;
 
-import com.nowayback.delivery.domain.exception.DeliveryDomainErrorCode;
-import com.nowayback.delivery.domain.exception.DeliveryDomainException;
+import com.nowayback.delivery.domain.delivery.exception.DeliveryDomainErrorCode;
+import com.nowayback.delivery.domain.delivery.exception.DeliveryDomainException;
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
