+++ delivery/src/main/java/com/nowayback/delivery/domain/deliverymanager/entity/DeliveryManager.java
@@ -4,8 +4,8 @@ import com.nowayback.common.audit.BaseEntity;
 import com.nowayback.delivery.domain.deliverymanager.vo.DeliveryManagerType;
 import com.nowayback.delivery.domain.deliverymanager.vo.DeliverySequence;
 import com.nowayback.delivery.domain.deliverymanager.vo.HubId;
-import com.nowayback.delivery.domain.exception.DeliveryManagerDomainErrorCode;
-import com.nowayback.delivery.domain.exception.DeliveryManagerDomainException;
+import com.nowayback.delivery.domain.deliverymanager.exception.DeliveryManagerDomainErrorCode;
+import com.nowayback.delivery.domain.deliverymanager.exception.DeliveryManagerDomainException;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
