+++ delivery/src/main/java/com/nowayback/delivery/domain/deliverymanager/exception/DeliveryManagerDomainException.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.domain.exception;
+package com.nowayback.delivery.domain.deliverymanager.exception;
 
 import com.nowayback.common.exception.GlobalException;
