+++ delivery/src/main/java/com/nowayback/delivery/domain/deliveryroute/exception/DeliveryRouteDomainErrorCode.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.domain.exception;
+package com.nowayback.delivery.domain.deliveryroute.exception;
 
 import com.nowayback.common.exception.ErrorCode;
 import org.springframework.http.HttpStatus;
