+++ delivery/src/test/java/com/nowayback/delivery/domain/deliverymanager/entity/DeliveryManagerTest.java
@@ -3,8 +3,8 @@ package com.nowayback.delivery.domain.deliverymanager.entity;
 import com.nowayback.delivery.domain.deliverymanager.vo.DeliveryManagerType;
 import com.nowayback.delivery.domain.deliverymanager.vo.DeliverySequence;
 import com.nowayback.delivery.domain.deliverymanager.vo.HubId;
-import com.nowayback.delivery.domain.exception.DeliveryManagerDomainErrorCode;
-import com.nowayback.delivery.domain.exception.DeliveryManagerDomainException;
+import com.nowayback.delivery.domain.deliverymanager.exception.DeliveryManagerDomainErrorCode;
+import com.nowayback.delivery.domain.deliverymanager.exception.DeliveryManagerDomainException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
